+++ Classes/MagiciteEffectFireBall.h
@@ -10,7 +10,7 @@ class MagiciteEffectFireBall : public MagiciteEffectShoot
 {
 public:
 
-    enum : int { cool_down = 3 };
+    enum : long long { cool_down = 3 };
 
     MagiciteEffectFireBall();
