+++ Classes/MagiciteEffectFlash.h
@@ -11,7 +11,7 @@ class MagiciteEffectFlash : public MagiciteEffectPositive
 {
 public:
 
-    enum : int { cool_down = 3 };
+    enum : long long { cool_down = 3 };
 
     MagiciteEffectFlash();
